ReadInt: Read numbers as (u)int64
Until now, ints have always been read as i32 which made it impossible to read
values higher than 2^31 even JSON allows it.
Now ints are first tried to be read as int64. If thats successfull and the read
number would fit into an int32 it's casted to an i32 as that data type is more
common.
If the number cannot be read as int64, it's tried to read it as uint64.
If that also fails, the number must be very big and the only way to read
it is by using a double. That may seem weird as the method is called 'ReadInt',
but JSON allows such big numbers so that's the best thing we could do.